Allows app to run locally reflecting the changes made to get app to deploy.
