+++ .gitignore
@@ -2,3 +2,4 @@
 .vscode
 .DS_Store
 mySQLCredentials.py
+app.egg-info
